+++ Package.swift
@@ -42,8 +42,7 @@ var dependencies: [Package.Dependency] = [
     .package(
         url: "https://github.com/pointfreeco/swift-snapshot-testing",
         revision: "26ed3a2b4a2df47917ca9b790a57f91285b923fb"
-    ),
-    .package(url: "https://github.com/airbnb/lottie-spm", from: "4.0.0")
+    )
 ]
 if shouldIncludeDocCPlugin {
     // Versions 1.4.0 and 1.4.1 are failing to compile, so we are pinning it to 1.3.0 for now
@@ -76,9 +75,7 @@ let package = Package(
         .library(name: "ReceiptParser",
                  targets: ["ReceiptParser"]),
         .library(name: "RevenueCatUI",
-                 targets: ["RevenueCatUI"]),
-        .library(name: "RevenueCatUILottie",
-                 targets: ["RevenueCatUILottie"])
+                 targets: ["RevenueCatUI"])
     ],
     dependencies: dependencies,
     targets: [
@@ -117,14 +114,6 @@ let package = Package(
                     .copy("Resources/background.jpg"),
                     .process("Resources/icons.xcassets")
                 ],
-                swiftSettings: ciCompilerFlags + additionalCompilerFlags),
-        // RevenueCatUILottie
-        .target(name: "RevenueCatUILottie",
-                dependencies: [
-                    "RevenueCatUI",
-                    .product(name: "Lottie", package: "lottie-spm")
-                ],
-                path: "RevenueCatUILottie",
                 swiftSettings: ciCompilerFlags + additionalCompilerFlags),
         .testTarget(name: "RevenueCatUITests",
                     dependencies: [
