+++ RevenueCatUI/Templates/V2/Components/Slot/SlotComponentView.swift
@@ -15,87 +15,6 @@ import Foundation
 import RevenueCat
 import SwiftUI
 
-private extension PaywallComponent {
-
-    var type: PaywallComponent.ComponentType {
-        switch self {
-        case .text:
-            return .text
-        case .image:
-            return .image
-        case .icon:
-            return .icon
-        case .stack:
-            return .stack
-        case .button:
-            return .button
-        case .package:
-            return .package
-        case .purchaseButton:
-            return .purchaseButton
-        case .stickyFooter:
-            return .stickyFooter
-        case .timeline:
-            return .timeline
-        case .tabs:
-            return .tabs
-        case .tabControl:
-            return .tabControl
-        case .tabControlButton:
-            return .tabControlButton
-        case .tabControlToggle:
-            return .tabControlToggle
-        case .carousel:
-            return .carousel
-        case .slot:
-            return .slot
-        case .slotLottie:
-            return .slotLottie
-        }
-    }
-
-}
-
-// Returns AnyView so we can store any SwiftUI view
-public typealias ViewProvider = (String) -> AnyView
-public typealias ViewOtherProvider = (PaywallComponent) -> AnyView
-
-public final class ViewRegistry: ObservableObject {
-
-    public static let shared = ViewRegistry()
-
-    private var viewCallback: ViewProvider?
-
-    private var otherCallback: [PaywallComponent.ComponentType: ViewOtherProvider] = [:]
-
-    internal init() {}
-
-    // Register a view-producing callback
-    public func register<Content: View>(type: PaywallComponent.ComponentType, @ViewBuilder _ callback: @escaping (PaywallComponent) -> Content) {
-
-        self.otherCallback[type] = { component in
-            AnyView(callback(component))
-        }
-    }
-
-    // Register a view-producing callback
-    public func register<Content: View>(@ViewBuilder _ callback: @escaping (String) -> Content) {
-        self.viewCallback = { id in
-            AnyView(callback(id))
-        }
-    }
-
-    // Build the view; safe fallback wrapped as AnyView
-    func makeView(identifier: String) -> AnyView {
-        viewCallback?(identifier) ?? AnyView(EmptyView())
-    }
-
-    func makeView(component: PaywallComponent) -> AnyView {
-        self.otherCallback[component.type]?(component) ?? AnyView(EmptyView())
-    }
-
-}
-
 #if !os(macOS) && !os(tvOS) // For Paywalls V2
 
 @available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
@@ -148,17 +67,14 @@ struct SlotComponentView: View {
 struct SlotComponentView_Previews: PreviewProvider {
 
     static let viewRegistry: ViewRegistry = {
-        let vr = ViewRegistry()
-
-        vr.register { identifier in
+        let viewRegistry = ViewRegistry()
+        viewRegistry.register { identifier in
             Text("Preview Slot \(identifier)")
         }
-
-        vr.register(type: .slotLottie) { component in
+        viewRegistry.register(type: .slotLottie) { _ in
             Text("Lottie goes here")
         }
-
-        return vr
+        return viewRegistry
     }()
 
     // Need to wrap in VStack otherwise preview rerenders and images won't show
@@ -186,6 +102,5 @@ struct SlotComponentView_Previews: PreviewProvider {
 }
 
 #endif
-
 
 #endif
