+++ RevenueCatUI/Templates/V2/Components/Slot/SlotComponentViewModel.swift
@@ -68,7 +68,6 @@ class SlotComponentViewModel {
 
 }
 
-
 extension PresentedSlotPartial: PresentedPartial {
 
     static func combine(
