--- RevenueCatUILottie/RevenueCatUILottie.swift
@@ -1,10 +1,0 @@
-//
-//  RevenueCatUILottie.swift
-//  RevenueCat
-//
-//  Created by Josh Holtz on 8/18/25.
-//
-
-public struct RevenueCatUILottie {
-
-}
