+++ Sources/Paywalls/Components/Common/PaywallComponentBase.swift
@@ -73,7 +73,7 @@ extension PaywallComponent {
 
     }
 
-    // swiftlint:disable:next cyclomatic_complexity
+    // swiftlint:disable:next cyclomatic_complexity function_body_length
     public func encode(to encoder: any Encoder) throws {
         var container = encoder.container(keyedBy: CodingKeys.self)
